Systematically eliminating compiler warnings.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@5117 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/PowerPC/JitCommon/JitCache.cpp
@@ -272,7 +272,7 @@ bool JitBlock::ContainsAddress(u32 em_address)
 #endif
 		if (inst & 0xfc000000) // definitely not a JIT block
 			return -1;
-		if (inst >= num_blocks)
+		if ((int)inst >= num_blocks)
 			return -1;
 		if (blocks[inst].originalAddress != addr)
 			return -1;
